+++ app/src/test/java/com/jim/sharetocomputer/RobolectricApplication.kt
@@ -17,20 +17,12 @@
 package com.jim.sharetocomputer
 
 import android.app.Application
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
-import android.net.NetworkInfo
 import com.elvishew.xlog.LogConfiguration
 import com.elvishew.xlog.LogLevel
 import com.elvishew.xlog.XLog
 import com.elvishew.xlog.printer.ConsolePrinter
 import com.jim.sharetocomputer.coroutines.DirectDispatcher
 import com.jim.sharetocomputer.coroutines.TestableDispatchers
-import org.robolectric.Shadows
-import org.robolectric.shadows.ShadowNetwork
-import org.robolectric.shadows.ShadowNetworkCapabilities
-import org.robolectric.shadows.ShadowNetworkInfo
 
 
 class RobolectricApplication: Application() {
@@ -44,27 +36,7 @@ class RobolectricApplication: Application() {
 
         WebServerService.isRunning.value = false
 
-        setupConnection()
-    }
-
-    @Suppress("DEPRECATION")
-    private fun setupConnection() {
-        val wifiNetworkInfo = ShadowNetworkInfo.newInstance(
-            NetworkInfo.DetailedState.CONNECTED,
-            ConnectivityManager.TYPE_WIFI,
-            0,
-            true,
-            NetworkInfo.State.CONNECTED
-        )
-        val network = ShadowNetwork.newInstance(ConnectivityManager.TYPE_WIFI)
-        val networkCapabilities = ShadowNetworkCapabilities.newInstance()
-        Shadows.shadowOf(networkCapabilities).addTransportType(NetworkCapabilities.TRANSPORT_WIFI)
-        val connectivityManager = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        Shadows.shadowOf(connectivityManager).apply {
-            addNetwork(network, wifiNetworkInfo)
-            setActiveNetworkInfo(wifiNetworkInfo)
-            setNetworkCapabilities(network, networkCapabilities)
-        }
+        setupWifiConnection()
     }
 
     private fun setupLogging() {
